Fixed Series/Readlist book dialog on Reader Screen Add going to next/prec comic in series after hitting last/first page in reader screen

+++ app/src/main/java/org/maddiesoftware/komagareader/komga_server/data/remote/dto/BookMetadataDto.kt
@@ -11,7 +11,7 @@ data class BookMetadataDto(
     val linksLock: Boolean?,
     val number: String?,
     val numberLock: Boolean?,
-    val numberSort: Int?,
+    val numberSort: Float?,
     val numberSortLock: Boolean?,
     val releaseDate: String?,
     val releaseDateLock: Boolean?,
